pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "client_project_template"
version = "0.1.0"
description = "Enterprise client project template using Kailash SDK"
authors = [
    {name = "Integrum", email = "info@integrum.com"}
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # Core Kailash SDK (from PyPI)
    "kailash>=0.6.6",
    
    # Web Framework
    "fastapi[all]>=0.115.0",
    "uvicorn[standard]>=0.24.0",
    
    # CLI Framework
    "click>=8.0",
    "rich>=13.0.0",
    
    # Data Handling
    "pydantic>=2.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    
    # Database
    "sqlalchemy>=2.0",
    "alembic>=1.12.0",
    
    # Development Tools
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
    
    # Utilities
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "black>=22.0",
    "isort>=5.10",
    "mypy>=0.9",
]

[project.urls]
"Homepage" = "https://github.com/integrum/new_project_template"
"Bug Tracker" = "https://github.com/integrum/new_project_template/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "solutions*"]
exclude = ["tests*", ".*", "__pycache__*"]

[dependency-groups]
dev = [
    "detect-secrets>=1.5.0",
    "doc8>=1.1.2",
    "pre-commit>=4.2.0",
]
